+++ DAL/DAL.csproj
@@ -58,8 +58,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="DatabaseContext.cs" />
-    <Compile Include="IDatabaseContext.cs" />
+    <Compile Include="Contexts\DatabaseContext.cs" />
+    <Compile Include="Contexts\IDatabaseContext.cs" />
     <Compile Include="Migrations\201910042248544_Initial.cs" />
     <Compile Include="Migrations\201910042248544_Initial.designer.cs">
       <DependentUpon>201910042248544_Initial.cs</DependentUpon>
@@ -67,7 +67,7 @@
     <Compile Include="Migrations\Configuration.cs" />
     <Compile Include="Seed\ICustomerSeed.cs" />
     <Compile Include="Seed\ITransactionSeed.cs" />
-    <Compile Include="TestDatabaseInitializer.cs" />
+    <Compile Include="Initializers\TestDatabaseInitializer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Seed\CustomerSeed.cs" />
     <Compile Include="Seed\TransactionSeed.cs" />
