+++ WebAPILab/Controllers/IHomeController.cs
@@ -1,5 +1,4 @@
-﻿using System.Threading.Tasks;
-using System.Web.Mvc;
+﻿using System.Web.Mvc;
 
 namespace WebAPILab.Controllers
 {
